Add files via upload
Changed Personality Insights to Senior Citizens Connect
Changed IBM Watson in the top left to maybe “Call for Code” or “St. Vrain Innovation Center” or “Silver Creek”
Added logos for the Innovation Center and Silver Creek in the top bar
Got rid of “Tweets and Replies” as well as “Your Twitter Personality” tabs
Changed terms of service to the one we will provide
Deleted the Choose section under Body of Text
Got rid of the “Resources” under Personality Insights
Changed the first paragraph under “Personality Insights” to our mission statement
Changed “Try the Service” to “Instructions” and change the paragraph under that to state what the seniors and students should do and how it works.